feat: add Error handler to manage errors in utils

+++ src/controllers/projects-controller.js
@@ -1,5 +1,6 @@
 const { Sequelize, Op } = require('sequelize')
 const { Project, Technology } = require('../db')
+const AppError = require('../utils');
 
 const getAllProjectsController = async (queries) => {
 	const { search, technologies, sort, page = 1, pageSize = 10 } = queries
@@ -39,19 +40,20 @@ const getAllProjectsController = async (queries) => {
 
 		return projects
 	} catch (error) {
-		console.error('Error fetching projects:', error)
-		throw new Error('Error fetching projects')
+        throw new AppError('Error fetching projects', 500);
 	}
 }
 
+
 const getProjectByIdController = async (id) => {
-	try {
-		const project = await Project.findByPk(id)
-		return project
-	} catch (error) {
-		throw error
-	}
-}
+    try {
+        const project = await Project.findByPk(id);
+        if (!project) throw new AppError(`Project with id ${id} not found`, 404);
+        return project;
+    } catch (error) {
+        throw new AppError(`Error fetching project with id ${id}`, 500);
+    }
+};
 
 const createProjectController = async (projectData, user) => {
 	try {
@@ -60,15 +62,15 @@ const createProjectController = async (projectData, user) => {
 			where: { title, userId: user.id },
 			defaults: { description, tags, image },
 		})
-		if (!created) throw new Error('This project already exists in DB!')
-		if (!technologies || technologies.length < 1) throw new Error('Add at least 1 technology')
+		if (!created) throw new AppError('This project already exists in the database!', 400);
+        if (!technologies || technologies.length < 1) throw new AppError('Add at least one technology', 400);
 
 		const techNames = technologies.map((tech) => (typeof tech === 'string' ? tech : tech.name))
 
 		const techInstances = await Technology.findAll({ where: { name: techNames } })
 
-		if (techInstances.length !== techNames.length)
-			throw new Error('Some technologies were not found in the DB')
+        if (techInstances.length !== techNames.length) 
+            throw new AppError('Some technologies were not found in the database', 400);
 
 		await project.addTechnologies(techInstances)
 
@@ -77,14 +79,14 @@ const createProjectController = async (projectData, user) => {
 			technologies: techNames,
 		}
 	} catch (error) {
-		throw new Error('Error creating a project', error)
+        throw new AppError('Error creating a project', 500);
 	}
 }
 
 const updateProjectController = async (projectData, id) => {
 	try {
-		const project = await Project.findByPk(id)
-		if (!project) throw new Error('Project not found')
+        const project = await Project.findByPk(id);
+        if (!project) throw new AppError('Project not found', 404);
 
 		await Project.update(
 			{
@@ -114,17 +116,18 @@ const updateProjectController = async (projectData, id) => {
 
 		return updatedProject
 	} catch (error) {
-		throw error
+        throw new AppError('Error updating project', 500);
 	}
 }
 
 const deleteProjectController = async (id) => {
 	try {
+		const project = await Project.findByPk(id);
+		if (!project) throw new AppError('Project not found', 404);
 		await Project.destroy({ where: { id: id } })
 		return 'Project correctly deleted'
 	} catch (error) {
-		throw error
-	}
+		throw new AppError('Error deleting project', 500);	}
 }
 
 module.exports = {
